+++ src/Responses/ErrorResponse.php
@@ -22,54 +22,62 @@
  * @copyright 2015 Datto, Inc.
  */
 
-namespace Datto\JsonRpc;
+namespace Datto\JsonRpc\Responses;
 
 /**
- * Class Error
- * @package Datto\JsonRpc
- *
  * A description of an error that occurred on the server
  *
- * @link http://www.jsonrpc.org/specification#error_object
+ * @link https://www.jsonrpc.org/specification#error_object
  */
-class Error
+class ErrorResponse extends Response
 {
+    const PARSE_ERROR = -32700;
+    const INVALID_ARGUMENTS = -32602;
+    const INVALID_METHOD = -32601;
+    const INVALID_REQUEST = -32600;
+
+    /** @var string */
     private $message;
 
+    /** @var int */
     private $code;
 
+    /** @var mixed */
     private $data;
 
     /**
-     * @param $message
-      * Short description of the error that occurred. This message SHOULD
-      * be limited to a single, concise sentence.
+     * @param mixed $id
+     * A unique identifier. This MUST be the same as the original request id.
+     * If there was an error while processing the request, then this MUST be null.
+     *
+     * @param string $message
+     * Short description of the error that occurred. This message SHOULD
+     * be limited to a single, concise sentence.
      *
      * @param int $code
      * Integer identifying the type of error that occurred.
      *
      * @param null|boolean|integer|float|string|array $data
-      * An optional primitive value that contains additional information about
+     * An optional primitive value that contains additional information about
      * the error.
      */
-    public function __construct($message, $code, $data)
+    public function __construct($id, string $message, int $code, $data = null)
     {
+        parent::__construct($id);
+
         $this->message = $message;
         $this->code = $code;
         $this->data = $data;
     }
 
-    /**
-     * @return mixed
-     */
-    public function getCode()
-    {
-        return $this->code;
-    }
-
-    public function getMessage()
+    public function getMessage(): string
     {
         return $this->message;
+    }
+
+    public function getCode(): int
+    {
+        return $this->code;
     }
 
     public function getData()